fix: extract shared skill utilities to eliminate DRY violations
Move duplicated parse_frontmatter() and find_skill_files() into a new
scripts/skill_utils.py module.  Both validate_plugins.py and
generate_marketplace.py now import from the shared module instead of
maintaining their own copies.

Closes #1110, Closes #928, Closes #914
